+++ src/app/settings/profile/layout.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function ProfileLayout({
+	children,
+}: {
+	children: React.ReactNode;
+}) {
+	return <section>{children}</section>;
+}
