+++ src/app/settings/security/layout.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function SecurityLayout({
+	children,
+}: {
+	children: React.ReactNode;
+}) {
+	return <section>{children}</section>;
+}
