+++ src/app/settings/security/page.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default async function Security() {
+	return <p>Security</p>;
+}
